pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "obsidian-rag-mcp"
version = "0.1.0"
description = "Semantic search for Obsidian vaults via MCP"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Claudio Garza", email = "theclaudiogarza@gmail.com" }
]
keywords = ["obsidian", "rag", "mcp", "semantic-search", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "chromadb>=0.4.0",
    "openai>=1.0.0",
    "mcp>=1.0.0",
    "python-frontmatter>=1.0.0",
    "mistune>=3.0.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "watchfiles>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
obsidian-rag = "src.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
